+++ src/migration.test.ts
@@ -1,7 +1,7 @@
 import { getBackendSrv } from '@grafana/runtime';
 
 import { getMigratedOptions } from '@/migration';
-import { CellType, ColumnEditorType, ColumnFilterMode, ColumnPinDirection, ImageScale, PermissionMode } from '@/types';
+import { CellType, ColumnEditorType, ColumnFilterMode, ColumnPinDirection, ImageScale, PaginationMode, PermissionMode } from '@/types';
 import {
   createColumnConfig,
   createColumnEditConfig,
@@ -10,6 +10,8 @@ import {
   createTableConfig,
   createTableRequestConfig,
 } from '@/utils';
+
+import { PAGE_SIZES } from './constants';
 
 /**
  * Mock @grafana/runtime
@@ -571,6 +573,33 @@ describe('migration', () => {
       expect(normalizedOptions.tables[0].items[3]).toEqual(
         expect.objectContaining({
           scale: ImageScale.PIXELATED,
+        })
+      );
+    });
+
+    it('Should normalize default page size', async () => {
+      const normalizedOptions = await getMigratedOptions({
+        pluginVersion: '1.8.0',
+        options: createPanelOptions({
+          tables: [
+            createTableConfig({
+              items: [
+                createColumnConfig({
+                  pin: true as never,
+                }),
+              ],
+              pagination: {
+                enabled: false,
+                mode: PaginationMode.CLIENT,
+              } as any,
+            }),
+          ],
+        }),
+      } as any);
+
+      expect(normalizedOptions.tables[0].pagination).toEqual(
+        expect.objectContaining({
+          defaultPageSize: PAGE_SIZES[0],
         })
       );
     });