+++ src/types/panel.ts
@@ -329,6 +329,13 @@ export interface TablePaginationConfig {
   mode: PaginationMode;
 
   /**
+   * Default page size
+   *
+   * @type {number}
+   */
+  defaultPageSize: number;
+
+  /**
    * Options for Query Mode
    */
   query?: {